[#2] refactor: 네이밍 오타 및 폴더구조 수정
