+++ src/features/main/league-adaptation/model/constants.ts
