+++ src/features/main/league-adaptation/model/mockData.ts
