+++ src/features/main/league-adaptation/model/type.ts
