+++ src/features/main/league-adaptation/ui/CtaBanner.tsx
