+++ src/features/main/league-adaptation/ui/LeagueAdaptationSection.tsx
