+++ src/features/main/league-adaptation/ui/LeagueSelector.tsx
