+++ src/pages/main/ui/FeatureHighlightSection.tsx
